Move away main thread access from app creation
