+++ autoconsent/autoconsent-impl/build.gradle
@@ -18,6 +18,7 @@ plugins {
     id 'com.android.library'
     id 'kotlin-android'
     id 'com.squareup.anvil'
+    id 'com.google.devtools.ksp' version "$ksp_version"
 }
 
 apply from: "$rootProject.projectDir/gradle/android-library.gradle"
@@ -41,7 +42,6 @@ dependencies {
     implementation project(path: ':app-build-config-api')
     implementation project(path: ':app-store')
     implementation project(path: ':autoconsent-api')
-    implementation project(path: ':autoconsent-store')
     implementation project(path: ':statistics')
     implementation project(path: ':browser-api')
     implementation project(path: ':navigation-api')
@@ -67,6 +67,9 @@ dependencies {
     implementation Square.moshi
     implementation "com.squareup.moshi:moshi-kotlin:_"
     implementation "com.squareup.moshi:moshi-adapters:_"
+    implementation AndroidX.room.runtime
+    implementation AndroidX.room.ktx
+    ksp AndroidX.room.compiler
 
     // Testing dependencies
     testImplementation project(path: ':common-test')
@@ -79,6 +82,7 @@ dependencies {
     testImplementation Testing.robolectric
     testImplementation 'app.cash.turbine:turbine:_'
     testImplementation project(path: ':common-test')
+    testImplementation AndroidX.room.testing
 
     coreLibraryDesugaring Android.tools.desugarJdkLibs
 }
