--- autoconsent/autoconsent-store/.gitignore
@@ -1,1 +1,0 @@
-/build
