--- autoconsent/autoconsent-store/build.gradle
@@ -1,38 +1,0 @@
-plugins {
-    id 'com.android.library'
-    id 'kotlin-android'
-    id 'com.google.devtools.ksp' version "$ksp_version"
-}
-
-apply from: "$rootProject.projectDir/gradle/android-library.gradle"
-
-dependencies {
-    implementation project(path: ':autoconsent-api')
-    implementation project(path: ':feature-toggles-api')
-    implementation project(path: ':common-utils')
-
-    implementation AndroidX.core.ktx
-    implementation KotlinX.coroutines.core
-
-    // Room
-    implementation AndroidX.room.runtime
-    implementation AndroidX.room.ktx
-    ksp AndroidX.room.compiler
-    testImplementation AndroidX.room.testing
-
-    // Testing dependencies
-    testImplementation Testing.junit4
-    testImplementation "org.mockito.kotlin:mockito-kotlin:_"
-    testImplementation AndroidX.test.ext.junit
-    testImplementation Testing.robolectric
-    testImplementation (KotlinX.coroutines.test) {
-        // https://github.com/Kotlin/kotlinx.coroutines/issues/2023
-        // conflicts with mockito due to direct inclusion of byte buddy
-        exclude group: "org.jetbrains.kotlinx", module: "kotlinx-coroutines-debug"
-    }
-    testImplementation project(path: ':common-test')
-}
-
-android {
-    namespace 'com.duckduckgo.autoconsent.store'
-}
